# gen/models/employeeofficial.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// Employeeofficial employeeofficial
//
// swagger:model employeeofficial
type Employeeofficial struct {

	// cnic
	// Example: 31202
	// Required: true
	Cnic *string `json:"cnic"`

	// department
	// Example: GoLang
	// Required: true
	Department *string `json:"department"`

	// designation
	// Example: BackEnd Developer
	// Required: true
	Designation *string `json:"designation"`

	// healthinsurance
	// Example: Nil
	// Required: true
	Healthinsurance *string `json:"healthinsurance"`

	// jobtype
	// Example: Intern
	// Required: true
	Jobtype *string `json:"jobtype"`

	// lifeinsurance
	// Example: Nil
	// Required: true
	Lifeinsurance *string `json:"lifeinsurance"`

	// salary
	// Example: 20000
	// Required: true
	Salary *string `json:"salary"`

	// teamlead
	// Example: Kashif Ali
	// Required: true
	Teamlead *string `json:"teamlead"`
}

// Validate validates this employeeofficial
func (m *Employeeofficial) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateCnic(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateDepartment(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateDesignation(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateHealthinsurance(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateJobtype(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateLifeinsurance(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateSalary(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateTeamlead(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *Employeeofficial) validateCnic(formats strfmt.Registry) error {

	if err := validate.Required("cnic", "body", m.Cnic); err != nil {
		return err
	}

	return nil
}

func (m *Employeeofficial) validateDepartment(formats strfmt.Registry) error {

	if err := validate.Required("department", "body", m.Department); err != nil {
		return err
	}

	return nil
}

func (m *Employeeofficial) validateDesignation(formats strfmt.Registry) error {

	if err := validate.Required("designation", "body", m.Designation); err != nil {
		return err
	}

	return nil
}

func (m *Employeeofficial) validateHealthinsurance(formats strfmt.Registry) error {

	if err := validate.Required("healthinsurance", "body", m.Healthinsurance); err != nil {
		return err
	}

	return nil
}

func (m *Employeeofficial) validateJobtype(formats strfmt.Registry) error {

	if err := validate.Required("jobtype", "body", m.Jobtype); err != nil {
		return err
	}

	return nil
}

func (m *Employeeofficial) validateLifeinsurance(formats strfmt.Registry) error {

	if err := validate.Required("lifeinsurance", "body", m.Lifeinsurance); err != nil {
		return err
	}

	return nil
}

func (m *Employeeofficial) validateSalary(formats strfmt.Registry) error {

	if err := validate.Required("salary", "body", m.Salary); err != nil {
		return err
	}

	return nil
}

func (m *Employeeofficial) validateTeamlead(formats strfmt.Registry) error {

	if err := validate.Required("teamlead", "body", m.Teamlead); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this employeeofficial based on context it is used
func (m *Employeeofficial) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *Employeeofficial) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Employeeofficial) UnmarshalBinary(b []byte) error {
	var res Employeeofficial
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}