# gen/restapi/operations/admin/getemployee_recoard_responses.go
// Code generated by go-swagger; DO NOT EDIT.

package admin

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"net/http"

	"github.com/go-openapi/runtime"

	"employeeManagementSystem/gen/models"
)

// GetemployeeRecoardOKCode is the HTTP code returned for type GetemployeeRecoardOK
const GetemployeeRecoardOKCode int = 200

/*GetemployeeRecoardOK successful operation

swagger:response getemployeeRecoardOK
*/
type GetemployeeRecoardOK struct {

	/*
	  In: Body
	*/
	Payload *models.Employeeofficial `json:"body,omitempty"`
}

// NewGetemployeeRecoardOK creates GetemployeeRecoardOK with default headers values
func NewGetemployeeRecoardOK() *GetemployeeRecoardOK {

	return &GetemployeeRecoardOK{}
}

// WithPayload adds the payload to the getemployee recoard o k response
func (o *GetemployeeRecoardOK) WithPayload(payload *models.Employeeofficial) *GetemployeeRecoardOK {
	o.Payload = payload
	return o
}

// SetPayload sets the payload to the getemployee recoard o k response
func (o *GetemployeeRecoardOK) SetPayload(payload *models.Employeeofficial) {
	o.Payload = payload
}

// WriteResponse to the client
func (o *GetemployeeRecoardOK) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.WriteHeader(200)
	if o.Payload != nil {
		payload := o.Payload
		if err := producer.Produce(rw, payload); err != nil {
			panic(err) // let the recovery middleware deal with this
		}
	}
}

// GetemployeeRecoardBadRequestCode is the HTTP code returned for type GetemployeeRecoardBadRequest
const GetemployeeRecoardBadRequestCode int = 400

/*GetemployeeRecoardBadRequest Invalid email supplied

swagger:response getemployeeRecoardBadRequest
*/
type GetemployeeRecoardBadRequest struct {
}

// NewGetemployeeRecoardBadRequest creates GetemployeeRecoardBadRequest with default headers values
func NewGetemployeeRecoardBadRequest() *GetemployeeRecoardBadRequest {

	return &GetemployeeRecoardBadRequest{}
}

// WriteResponse to the client
func (o *GetemployeeRecoardBadRequest) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(400)
}

// GetemployeeRecoardNotFoundCode is the HTTP code returned for type GetemployeeRecoardNotFound
const GetemployeeRecoardNotFoundCode int = 404

/*GetemployeeRecoardNotFound email not found

swagger:response getemployeeRecoardNotFound
*/
type GetemployeeRecoardNotFound struct {
}

// NewGetemployeeRecoardNotFound creates GetemployeeRecoardNotFound with default headers values
func NewGetemployeeRecoardNotFound() *GetemployeeRecoardNotFound {

	return &GetemployeeRecoardNotFound{}
}

// WriteResponse to the client
func (o *GetemployeeRecoardNotFound) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(404)
}