# swagger.yml
swagger: "2.0"
info:
  description: "This Microservice used to add the official information in the Employee Account"
  version: "1.0.0"
  title: "Employee Managemnt System"
  contact:
    email: "abdulmueez@wanclouds.net"
host: "ems.swagger.io"
basePath: "/v1"
...
schemes:
  - "http"
paths:
...
  /Admin/getrecoard/{cnic}:
    get:
      tags:
        - "Admin"
      summary: "Show Employee recoard"
      operationId: "getemployee recoard"
      produces:
        - "application/json"
      parameters:
        - name: "cnic"
          in: "path"
          description: "Enter CNIC to Show the Recoard"
          required: true
          type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/employeeofficial"
        "400":
          description: "Invalid email supplied"
        "404":
          description: "email not found"
...
definitions:
  employeeofficial:
    type: "object"
    required:
      - "cnic"
      - "salary"
      - "designation"
      - "department"
      - "teamlead"
      - "jobtype"
      - "healthinsurance"
      - "lifeinsurance"
    properties:
      cnic:
        type: "string"
        example: "31202"
      salary:
        type: "string"
        example: "20000"
      designation:
        type: "string"
        example: "BackEnd Developer"
      department:
        type: "string"
        example: "GoLang"
      teamlead:
        type: "string"
        example: "Kashif Ali"
      jobtype:
        type: "string"
        example: "Intern"
      healthinsurance:
        type: "string"
        example: "Nil"
      lifeinsurance:
        type: "string"
        example: "Nil"
...